pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zero-autogen-demo"
version = "0.1.0"
description = "CLI MVP: AutoGen agent autonomously builds schema and SQL on TiDB Zero"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "autogen-agentchat>=0.4.0",
  "autogen-ext[openai]>=0.4.0",
  "certifi>=2024.7.4",
  "httpx>=0.27.0",
  "pymysql>=1.1.1",
  "python-dotenv>=1.0.1",
  "typer>=0.12.5",
]

[project.scripts]
zero-agent-demo = "zero_autogen_demo.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
